no auth perfected + access in dev mode added

+++ pages/api/post/[id].ts
@@ -1,12 +1,13 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
 import prisma from '../../../lib/prisma'
+import {inDevEnvironment} from '../../../lib/DevEnv';
 
 
 // DELETE /api/post/:id
 export default async function handle(req: NextApiRequest, res: NextApiResponse) {
   const postId = req.query.id;
 
-  if (req.method === "DELETE") {
+  if (req.method === "DELETE" && inDevEnvironment) {
 
       const post = await prisma.post.delete({
         where: { id: Number(postId) },
